added "registration_type_cmp_manager" and fixed compile error
